+++ src/main/java/ru/otus/architect/command/Command.java
@@ -1,0 +1,6 @@
+package ru.otus.architect.command;
+
+public interface Command {
+    void execute();
+
+}
